backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.6.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cobble==0.1.4
coverage==7.8.0
cryptography==44.0.1
cssselect2==0.7.0
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
et_xmlfile==2.0.0
Faker==25.9.1
fire==0.7.0
fonttools==4.56.0
html5lib==1.1
idna==3.7
iniconfig==2.1.0
lxml==5.3.1
numpy==2.2.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.8.0
PyMuPDF==1.25.3
pypdf==5.3.0
PyPDF2==3.0.1
pyphen==0.17.2
pytest==8.3.5
pytest-cov==6.1.0
pytest-django==4.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
svglib==1.5.1
termcolor==2.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.17
XlsxWriter==3.2.0
zopfli==0.2.3.post1

# Security and validation dependencies
python-magic==0.4.27
